requirements.txt
fastapi[all]

google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib